Went ahead and did the reorganization before vm.go got too bad. Config specific methods moved to vmconfig.go (formerly  config.go) which had been empty for a while. The new fluent methods dropped code coverage down to roughly 84%, which is below the threshold, so won't be pushing to repo until back to normal.
