--- pkg/wasmvm/config.go
@@ -1,3 +1,0 @@
-package wasmvm
-
-// You can expand this later for more detailed config loading and validation.
